+++ src/interfaces/ViewOptions.ts
@@ -41,7 +41,7 @@ export default interface {
    * @example
    * // The message set here can be referenced in the view as {{message}}.
    * beforeRender: res => {
-   *   res.loclas.message = 'Hello World';
+   *   res.locals.message = 'Hello World';
    * }
    *
    * @type {(res: express.Response) => void}